requirements.txt
jieba==0.42.1
pinyin==0.4.0
